workload: add a racing game workload
BeachBuggyBlitz is a racing game with repeatable scene.

At the begging screen, it has a camera from 3-rd person angle rotating at
a fixed speed. So that the scenario is periodical and reproducible and it
also contains various type of scene, like ocean, waterfall. Those various
scens can be used to test how fast the dvfs governor adjusts to the proper
frequency.